+++ changelog.md
@@ -2,7 +2,13 @@
 
 ## 0.4
 
-- Added ability to add and receive commands for HomeTemp commanding from external
+This version adds the ability to create, receive and execute commands from external.
+- Updated `EmailDistributor` to support CRUD for emails
+- Added `Command`, `CommandRequest`, `CommandParser` and `CommandService` to enable commanding
+- Added `command_test.py` as (test) demo for commanding
+- Restructured and added new fields in `default_hometemp.ini/distribution`
+
+- Updated readme.md
 
 
 ## 0.3.3